# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TAU Service API",
    "description": "TAU Service Template API documentation",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "API v1"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
    "/users/me": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get current user",
        "description": "Returns the currently authenticated user's profile",
        "operationId": "getCurrentUser",
        "responses": {
          "200": {
            "description": "User profile retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetUserResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/platforms": {
      "get": {
        "tags": [
          "Platforms"
        ],
        "summary": "List platforms",
        "operationId": "listPlatforms",
        "description": "Returns the whole platform library, which is global rather than per-squadron. Open to any authenticated user.",
        "responses": {
          "200": {
            "description": "Platforms retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListPlatformsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Platforms"
        ],
        "summary": "Create a platform",
        "operationId": "createPlatform",
        "description": "Requires admin, editor, rto, or planner - PACE's writer set, because platforms are joint planning reference data.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePlatformRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Platform created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatePlatformResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed, or an abbrev the waveform library does not declare (PLATFORM_UNKNOWN_WAVEFORM), or a category or kind outside the vocabulary shape",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden. The caller is authenticated but lacks the role this route requires.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Another platform already uses this designation (PLATFORM_DESIGNATION_EXISTS)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/platforms/{id}": {
      "patch": {
        "tags": [
          "Platforms"
        ],
        "summary": "Update a platform",
        "operationId": "updatePlatform",
        "description": "Partial update; omitted fields are left as they are. Requires admin, editor, rto, or planner - PACE's writer set, because platforms are joint planning reference data.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Platform ID",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePlatformRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Platform updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdatePlatformResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed, or an abbrev the waveform library does not declare (PLATFORM_UNKNOWN_WAVEFORM)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden. The caller is authenticated but lacks the role this route requires.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Platform not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Another platform already uses this designation (PLATFORM_DESIGNATION_EXISTS)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Platforms"
        ],
        "summary": "Delete a platform",
        "operationId": "deletePlatform",
        "description": "Requires admin, editor, rto, or planner - PACE's writer set, because platforms are joint planning reference data.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Platform ID",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Platform deleted"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden. The caller is authenticated but lacks the role this route requires.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Platform not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/platforms/import": {
      "post": {
        "tags": [
          "Platforms"
        ],
        "summary": "Import platforms from CSV",
        "operationId": "importPlatforms",
        "description": "Bulk-imports platforms from a CSV payload. Returns the imported count and any per-row errors; a row naming an abbrev the waveform library does not declare fails. Requires admin, editor, rto, or planner - PACE's writer set, because platforms are joint planning reference data.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CsvImportRequest"
              }
            }
          }
        },
        "responses": {
          "207": {
            "description": "Import processed - some or all rows were written, with any per-row errors listed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CsvImportResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body, empty CSV, or every row failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden. The caller is authenticated but lacks the role this route requires.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/export/platforms": {
      "get": {
        "tags": [
          "Platforms"
        ],
        "summary": "Export platforms as CSV",
        "operationId": "exportPlatforms",
        "description": "Streams a CSV of platforms; the two arrays serialise as one ;-separated cell each. Read-only; open to any authenticated user.",
        "parameters": [
          {
            "name": "columns",
            "in": "query",
            "description": "Comma-separated column subset (empty = all, canonical order)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "CSV file returned",
            "content": {
              "text/csv": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT token"
      }
    },
    "schemas": {
      "UserResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "User's unique identifier"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "User's email address"
          },
          "name": {
            "type": "string",
            "description": "User's display name"
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "User's assigned roles"
          },
          "preferences": {
            "$ref": "#/components/schemas/PreferencesResponse"
          },
          "last_login_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "Last login timestamp"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Account creation timestamp"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "Last update timestamp"
          }
        },
        "required": [
          "id",
          "email",
          "name",
          "roles",
          "preferences",
          "created_at",
          "updated_at"
        ]
      },
      "PreferencesResponse": {
        "type": "object",
        "properties": {
          "theme": {
            "type": "string",
            "enum": [
              "light",
              "dark"
            ],
            "description": "User's preferred theme"
          }
        },
        "required": [
          "theme"
        ]
      },
      "GetUserResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object",
            "properties": {
              "user": {
                "$ref": "#/components/schemas/UserResponse"
              }
            }
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "description": "Error code for programmatic handling"
              },
              "message": {
                "type": "string",
                "description": "Human-readable error message"
              },
              "details": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ValidationError"
                },
                "description": "Validation error details (only present for validation errors)"
              }
            },
            "required": [
              "code",
              "message"
            ]
          }
        }
      },
      "ValidationError": {
        "type": "object",
        "properties": {
          "field": {
            "type": "string",
            "description": "Field that failed validation"
          },
          "message": {
            "type": "string",
            "description": "Validation error message"
          }
        },
        "required": [
          "field",
          "message"
        ]
      },
      "ImportRowError": {
        "type": "object",
        "properties": {
          "row": {
            "type": "integer",
            "description": "1-based CSV row number"
          },
          "name": {
            "type": "string",
            "description": "Name value from the offending row"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Validation errors for this row"
          }
        },
        "required": [
          "row",
          "name",
          "errors"
        ]
      },
...
      "CsvImportRequest": {
        "type": "object",
        "description": "The CSV import body shared by every domain that imports through csvhttp: equipment, waveforms, services, transports and nets. Terminals and kits predate it and keep their own request schemas.",
        "properties": {
          "csv": {
            "type": "string",
            "description": "Raw CSV content to import"
          }
        },
        "required": [
          "csv"
        ]
      },
      "CsvImportResponse": {
        "type": "object",
        "description": "The import result shared by every domain that imports through csvhttp. 207 whenever anything was written, including a completely clean import; 400 only when every row failed.",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object",
            "properties": {
              "imported": {
                "type": "integer",
                "description": "Number of rows written"
              },
              "errors": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ImportRowError"
                }
              },
              "message": {
                "type": "string",
                "description": "Human-readable import summary"
              }
            }
          }
        }
      },
...
      "PlatformResponse": {
        "type": "object",
        "description": "An external comms platform. Its waveform set lives in one row: no join table.",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "designation": {
            "type": "string",
            "maxLength": 60,
            "description": "Natural key (F-35A). Unique, case-insensitive."
          },
          "popular_name": {
            "type": "string",
            "maxLength": 120
          },
          "category": {
            "type": "string",
            "description": "Default joint. Suggestions: organic, joint, coalition. Open vocabulary, normalised to lowercase, trimmed, internal whitespace collapsed, underscores to spaces; anything outside [a-z0-9 -] up to 40 characters is refused with PLATFORM_INVALID_CATEGORY or PLATFORM_INVALID_KIND."
          },
          "kind": {
            "type": "string",
            "description": "Default aircraft. Suggestions: aircraft, ship, ground vehicle, ground station. Open vocabulary, normalised to lowercase, trimmed, internal whitespace collapsed, underscores to spaces; anything outside [a-z0-9 -] up to 40 characters is refused with PLATFORM_INVALID_CATEGORY or PLATFORM_INVALID_KIND."
          },
          "operator": {
            "type": "string",
            "maxLength": 120,
            "description": "Service or nation"
          },
          "waveform_abbrevs": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Waveforms declared on the platform directly. Each must exist in the waveform library; de-duplicated case-insensitively, keeping the first spelling."
          },
          "equipment_ids": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Catalog radios it carries, where known. An id naming a deleted record is kept and ignored by the matrix."
          },
          "notes": {
            "type": "string",
            "maxLength": 1000
          },
          "created_by": {
            "type": "string"
          },
          "updated_by": {
            "type": "string"
          },
          "created_at": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "designation",
          "popular_name",
          "category",
          "kind",
          "operator",
          "waveform_abbrevs",
          "equipment_ids",
          "notes"
        ]
      },
      "ListPlatformsResponse": {
        "type": "object",
        "properties": {
          "platforms": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlatformResponse"
            }
          },
          "total": {
            "type": "integer"
          }
        },
        "required": [
          "platforms",
          "total"
        ]
      },
      "CreatePlatformRequest": {
        "type": "object",
        "description": "Uniqueness is case-insensitive on designation. An abbrev the waveform library does not declare is refused with PLATFORM_UNKNOWN_WAVEFORM.",
        "required": [
          "designation"
        ],
        "properties": {
          "designation": {
            "type": "string",
            "minLength": 1,
            "maxLength": 60
          },
          "popular_name": {
            "type": "string",
            "maxLength": 120
          },
          "category": {
            "type": "string",
            "description": "Open vocabulary, normalised to lowercase, trimmed, internal whitespace collapsed, underscores to spaces; anything outside [a-z0-9 -] up to 40 characters is refused with PLATFORM_INVALID_CATEGORY or PLATFORM_INVALID_KIND."
          },
          "kind": {
            "type": "string",
            "description": "Open vocabulary, normalised to lowercase, trimmed, internal whitespace collapsed, underscores to spaces; anything outside [a-z0-9 -] up to 40 characters is refused with PLATFORM_INVALID_CATEGORY or PLATFORM_INVALID_KIND."
          },
          "operator": {
            "type": "string",
            "maxLength": 120
          },
          "waveform_abbrevs": {
            "type": "array",
            "items": {
              "type": "string",
              "maxLength": 40
            },
            "maxItems": 64
          },
          "equipment_ids": {
            "type": "array",
            "items": {
              "type": "string",
              "maxLength": 64
            },
            "maxItems": 64
          },
          "notes": {
            "type": "string",
            "maxLength": 1000
          }
        }
      },
      "UpdatePlatformRequest": {
        "type": "object",
        "description": "PATCH is a partial update: an omitted field is left as it is, which is why every property here is optional. A present waveform_abbrevs replaces the whole set.",
        "properties": {
          "designation": {
            "type": "string",
            "minLength": 1,
            "maxLength": 60
          },
          "popular_name": {
            "type": "string",
            "maxLength": 120
          },
          "category": {
            "type": "string",
            "description": "Open vocabulary, normalised to lowercase, trimmed, internal whitespace collapsed, underscores to spaces; anything outside [a-z0-9 -] up to 40 characters is refused with PLATFORM_INVALID_CATEGORY or PLATFORM_INVALID_KIND."
          },
          "kind": {
            "type": "string",
            "description": "Open vocabulary, normalised to lowercase, trimmed, internal whitespace collapsed, underscores to spaces; anything outside [a-z0-9 -] up to 40 characters is refused with PLATFORM_INVALID_CATEGORY or PLATFORM_INVALID_KIND."
          },
          "operator": {
            "type": "string",
            "maxLength": 120
          },
          "waveform_abbrevs": {
            "type": "array",
            "items": {
              "type": "string",
              "maxLength": 40
            },
            "maxItems": 64
          },
          "equipment_ids": {
            "type": "array",
            "items": {
              "type": "string",
              "maxLength": 64
            },
            "maxItems": 64
          },
          "notes": {
            "type": "string",
            "maxLength": 1000
          }
        }
      },
      "CreatePlatformResponse": {
        "type": "object",
        "properties": {
          "platform": {
            "$ref": "#/components/schemas/PlatformResponse"
          }
        },
        "required": [
          "platform"
        ]
      },
      "UpdatePlatformResponse": {
        "type": "object",
        "properties": {
          "platform": {
            "$ref": "#/components/schemas/PlatformResponse"
          }
        },
        "required": [
          "platform"
        ]
      }
    }
  },
...
}